[classes](chore) Reorganize compression class
